\name{sp500}
\docType{data}
\alias{sp500}
\title{Example options of the S&P 500}
\description{
  This data set covers option data for the S&P 500 as observed on ....
}
\usage{data(sp500)}
\format{A list of the options with different maturities and strikes.}
\source{OptionMetrics.}
\references{
  ...
}
\keyword{datasets}
